% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction-helpers.R
\name{begin_transaction}
\alias{begin_transaction}
\title{Begin a Database Transaction}
\usage{
begin_transaction(con)
}
\arguments{
\item{con}{A DBI connection object.}
}
\description{
This function starts a transaction by executing the \code{BEGIN} statement.
}
\keyword{internal}
