% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-definition.R
\name{calculate_char_size}
\alias{calculate_char_size}
\title{Calculate Character Size for Column}
\usage{
calculate_char_size(col)
}
\arguments{
\item{col}{A column from a data frame.}
}
\value{
An integer representing the maximum character size for the column.
}
\description{
This function calculates the appropriate character size for a given column in the data frame,
which is used to determine the VARCHAR size for Redshift table creation.
}
\keyword{internal}
