% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-definition.R
\name{col_to_redshift_type}
\alias{col_to_redshift_type}
\title{Convert Column to Redshift Type}
\usage{
col_to_redshift_type(col, compression)
}
\arguments{
\item{col}{A column from a data frame.}

\item{compression}{A logical value indicating whether compression should be applied.}
}
\value{
A character string representing the Redshift data type for the column.
}
\description{
This function determines the appropriate Redshift data type for a given column in the data frame.
}
\keyword{internal}
