% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction-helpers.R
\name{commit_transaction}
\alias{commit_transaction}
\title{Commit a Database Transaction}
\usage{
commit_transaction(con)
}
\arguments{
\item{con}{A DBI connection object.}
}
\description{
This function commits a transaction by executing the \code{COMMIT} statement.
}
\keyword{internal}
