% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{deletePrefix}
\alias{deletePrefix}
\title{Delete S3 Objects with a Given Prefix}
\usage{
deletePrefix(prefix, bucket, split_files, key, secret, session, region)
}
\arguments{
\item{prefix}{The prefix of the objects to delete in the S3 bucket.}

\item{bucket}{The name of the S3 bucket.}

\item{split_files}{Number of files to delete.}

\item{key}{AWS access key for authentication.}

\item{secret}{AWS secret key for authentication.}

\item{session}{AWS session token for authentication (optional).}

\item{region}{AWS region where the bucket is located.}
}
\description{
This function deletes multiple objects in an S3 bucket that share a common prefix.
It retries the deletion process in case of 500 errors.
}
\keyword{internal}
