% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_utils.R
\name{list_tables}
\alias{list_tables}
\title{List Redshift tables}
\usage{
list_tables(db_con)
}
\arguments{
\item{db_con}{a DBI connection to the Redshift server}
}
\value{
a data.frame of table names
}
\description{
Lists all tables available in Redshift database
}
