% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{queryDo}
\alias{queryDo}
\title{Execute a SQL Query and Return Results}
\usage{
queryDo(dbcon, query)
}
\arguments{
\item{dbcon}{A database connection object.}

\item{query}{A SQL query string to be executed.}
}
\value{
A data frame containing the query results.
}
\description{
This function executes a SQL query and returns the results.
}
\keyword{internal}
