% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{queryStmt}
\alias{queryStmt}
\title{Execute a SQL Statement Without Returning Results}
\usage{
queryStmt(dbcon, query)
}
\arguments{
\item{dbcon}{A database connection object.}

\item{query}{A SQL statement string to be executed.}
}
\description{
This function executes a SQL statement that does not return results (e.g., CREATE, INSERT).
}
\keyword{internal}
