% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftr-package.R
\docType{package}
\name{redshiftr-package}
\alias{redshiftr}
\alias{redshiftr-package}
\title{redshiftr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
