% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction-helpers.R
\name{rollback_transaction}
\alias{rollback_transaction}
\title{Rollback a Database Transaction}
\usage{
rollback_transaction(con)
}
\arguments{
\item{con}{A DBI connection object.}
}
\description{
This function rolls back a transaction by executing the \code{ROLLBACK} statement.
}
\keyword{internal}
