% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-definition.R
\name{rs_create_statement}
\alias{rs_create_statement}
\title{Create Redshift Table Statement}
\usage{
rs_create_statement(
  df,
  table_name,
  sortkeys = NULL,
  sortkey_style = "compound",
  distkey = NULL,
  distkey_style = "even",
  compression = TRUE
)
}
\arguments{
\item{df}{A data frame to use for generating the CREATE TABLE statement.}

\item{table_name}{A character string representing the name of the table to be created.}

\item{sortkeys}{A character vector of column names to be used as sort keys (optional).}

\item{sortkey_style}{A character string specifying the sort key style, either 'compound' or 'interleaved' (default is 'compound').}

\item{distkey}{A character string representing the distribution key column (optional).}

\item{distkey_style}{A character string specifying the distribution style, either 'even' or 'all' (default is 'even').}

\item{compression}{A logical value indicating whether compression should be applied (default is TRUE).}
}
\value{
A character string representing the CREATE TABLE statement for Redshift.
}
\description{
This function generates a SQL CREATE TABLE statement for Redshift based on the provided data frame and parameters.
}
\keyword{internal}
