% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rs_delete_table}
\alias{rs_delete_table}
\title{Copy Data from S3 to Redshift}
\usage{
rs_delete_table(con, table_name)
}
\arguments{
\item{con}{A DBI connection object to Redshift.}

\item{table_name}{The name of the table to be deleted.}
}
\value{
NULL. The table is deleted from Redshift.
}
\description{
This function copies data from an S3 bucket to a Redshift table using the COPY command.
It
}
\details{
Delete a table from Redshift

This function generates and executes a \verb{DROP TABLE} SQL statement to delete a table from Redshift.
}
