% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-sql.R
\name{rs_insert_sql}
\alias{rs_insert_sql}
\title{Generate a single SQL insert command for multiple rows}
\usage{
rs_insert_sql(df_chunk, table_name, con)
}
\arguments{
\item{df_chunk}{A chunk of the data frame for which to generate a single \verb{INSERT INTO} statement.}

\item{table_name}{The name of the table where data will be inserted.}

\item{con}{A DBI connection object to Redshift.}
}
\value{
A single string containing a combined SQL \verb{INSERT INTO} statement.
}
\description{
This function generates a combined SQL \verb{INSERT INTO} statement for a chunk of data,
so that all rows in the chunk are inserted with one \code{dbExecute()} call.
}
