% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_utils.R
\name{show_table_info}
\alias{show_table_info}
\title{Show table information}
\usage{
show_table_info(db_con, table_name, schema_name = "public")
}
\arguments{
\item{db_con}{a DBI connection to the Redshift server}

\item{table_name}{table name}

\item{schema_name}{schema name, defaults to 'public'}
}
\value{
a data.frame containing table information
}
\description{
Show table information
}
