% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{splitDetermine}
\alias{splitDetermine}
\title{Determine Optimal Split Size for Data Upload}
\usage{
splitDetermine(dbcon, numRows, rowSize)
}
\arguments{
\item{dbcon}{A database connection object.}

\item{numRows}{Number of rows in the data.}

\item{rowSize}{Estimated size of each row in bytes.}
}
\value{
The number of files to split the data into.
}
\description{
This function determines the optimal number of files to split the data into for uploading to Redshift.
It calculates the number of slices in Redshift and adjusts the split size accordingly.
}
\keyword{internal}
