% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{uploadToS3}
\alias{uploadToS3}
\title{Upload Data to S3 Bucket}
\usage{
uploadToS3(data, bucket, split_files, key, secret, session, region)
}
\arguments{
\item{data}{A data frame to upload to S3.}

\item{bucket}{The name of the S3 bucket.}

\item{split_files}{Number of files to split the data into for upload.}

\item{key}{AWS access key for authentication.}

\item{secret}{AWS secret key for authentication.}

\item{session}{AWS session token for authentication (optional).}

\item{region}{AWS region where the bucket is located.}
}
\value{
A prefix string used for the uploaded files, or NA if the upload fails.
}
\description{
This function uploads a data frame to an S3 bucket, splitting it into multiple files if necessary.
The function checks if the bucket exists and retries in case of errors during upload.
}
\keyword{internal}
